+++ core/data/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/data/service/ResultService.kt
@@ -6,8 +6,8 @@ import kotlinx.coroutines.flow.Flow
 
 interface ResultService {
     fun getFeedbackDetailsForResult(
-        participationId: Int,
-        resultId: Int,
+        participationId: Long,
+        resultId: Long,
         serverUrl: String,
         bearerToken: String
     ): Flow<DataState<List<Feedback>>>
