+++ core/data/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/data/service/impl/ResultServiceImpl.kt
@@ -16,8 +16,8 @@ internal class ResultServiceImpl(
     private val ktorProvider: KtorProvider
 ) : ResultService {
     override fun getFeedbackDetailsForResult(
-        participationId: Int,
-        resultId: Int,
+        participationId: Long,
+        resultId: Long,
         serverUrl: String,
         bearerToken: String
     ): Flow<DataState<List<Feedback>>> {
