+++ core/model/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/model/exercise/participation/StudentParticipation.kt
@@ -23,7 +23,7 @@ abstract class StudentParticipation : Participation() {
     @Serializable
     @SerialName("student")
     class StudentParticipationImpl(
-        override val id: Int? = null,
+        override val id: Long? = null,
         override val initializationState: InitializationState? = null,
         override val initializationDate: Instant? = null,
         override val individualDueDate: Instant? = null,
